- Added network sync in tank

+++ P2PTank/P2PTank/SharedSource/Main/Entities/P2PMessages/MoveMessage.cs
@@ -3,7 +3,9 @@
     public class MoveMessage
     {
         public string PlayerId { get; set; }
-        public double X { get; set; }
-        public double Y { get; set; }
+
+        public float X { get; set; }
+
+        public float Y { get; set; }
     }
 }
