+++ P2PTank/P2PTank/SharedSource/Main/Entities/P2PMessages/RotateMessage.cs
@@ -3,6 +3,7 @@
     public class RotateMessage
     {
         public string PlayerId { get; set; }
-        public double Rotation { get; set; }
+
+        public float Rotation { get; set; }
     }
 }
